refactor: FormDialog と別概念なため story の名称変更
